+++ src/nsidc/icesat2gis/atl08.py
@@ -10,7 +10,7 @@ from pyproj import Geod
 from shapely import LineString, Point
 from shapely.geometry import MultiLineString
 
-from nsidc.icesat2gis.exceptions import IceSatMissingDataError
+from nsidc.icesat2gis.exceptions import IceSat2MissingDataError
 
 GroundTrack = Literal["gt1l", "gt1r", "gt2l", "gt2r", "gt3l", "gt3r"]
 
@@ -18,13 +18,20 @@ GroundTrack = Literal["gt1l", "gt1r", "gt2l", "gt2r", "gt3l", "gt3r"]
 # Default ground_track core variables for ATL08.
 ATL08_DEFAULT_GT_CORE_VARS = (
     "canopy/h_canopy",
+    "canopy/h_mean_canopy",
     "canopy/h_canopy_uncertainty",
     "canopy/h_median_canopy",
     "canopy/photon_rate_can",
     "terrain/h_te_best_fit",
+    "terrain/h_te_mean",
     "terrain/h_te_uncertainty",
     "terrain/photon_rate_te",
     "terrain/terrain_slope",
+)
+
+ATL08_DEFAULT_VARIABLES_TO_CHECK_ALL_NULL = (
+    "canopy/h_canopy",
+    "terrain/h_te_best_fit",
 )
 
 
@@ -33,13 +40,32 @@ def _read_points_for_gt(
     ground_track: GroundTrack,
     filepath: Path,
     variables_to_include: Sequence[str],
+    variables_to_check_all_null: Sequence[str],
 ) -> gpd.GeoDataFrame:
     """Reads 100m segment points from ATL08 for the given ground track.
 
-    Raises an `IceSatMissingDataError` when a ground track is missing data
+    Raises an `IceSat2MissingDataError` when a ground track is missing data
     (either the ground track group or the ground track's `land_segments` group
     is missing).
+
+    * `ground_track`: one of "gt1l", "gt1r", "gt2l", "gt2r", "gt3l", "gt3r"
+    * `filepath`: filepath to the ATL08 granule
+    * `variables_to_include`: sequence of strings representing GT variables to
+      include in the output (e.g,. `"canopy/h_canopy"`).
+    * `variables_to_check_all_null`: sequence of strings representing GT
+      variables to check for Null values. If all of the variables specified in
+      this sequence are Null for a record, that record is excluded.
     """
+    if not all(
+        var_to_check in variables_to_include
+        for var_to_check in variables_to_check_all_null
+    ):
+        msg = (
+            f"All `variables_to_check_all_null` must be in `variables_to_include`."
+            f"Got {variables_to_include=} {variables_to_check_all_null=}."
+        )
+        raise ValueError(msg)
+
     try:
         ds = xr.open_datatree(
             filepath,
@@ -51,7 +77,7 @@ def _read_points_for_gt(
         # `land_segments` group is missing.
         msg = f"No `land_segment` data for {ground_track} from {filepath}: {e}"
         print(msg)
-        raise IceSatMissingDataError(msg) from e
+        raise IceSat2MissingDataError(msg) from e
 
     # Extract variables
     lats = ds.latitude
@@ -77,6 +103,16 @@ def _read_points_for_gt(
         crs="EPSG:4326",
     )
 
+    # Drop points that are all-NaN for the user's selected variables.
+    if variables_to_check_all_null:
+        variables_to_check_all_null_names = [
+            var.rsplit("/", maxsplit=1)[-1] for var in variables_to_check_all_null
+        ]
+        gdf = gdf.dropna(
+            subset=variables_to_check_all_null_names,
+            how="all",
+        ).reset_index()
+
     # Localize the timestamp to UTC. Otherwise it inherits the system TZ
     # (e.g., MST).
     gdf["delta_time"] = gdf.delta_time.dt.tz_localize("UTC")
@@ -88,6 +124,9 @@ def read_points_from_atl08(
     *,
     filepath: Path,
     gt_variables_to_include: Sequence[str] = ATL08_DEFAULT_GT_CORE_VARS,
+    gt_variables_to_check_all_null: Sequence[
+        str
+    ] = ATL08_DEFAULT_VARIABLES_TO_CHECK_ALL_NULL,
 ) -> gpd.GeoDataFrame:
     """Return a GeoDataFrame containing points representing ground tracks."""
     gdfs = []
@@ -97,14 +136,15 @@ def read_points_from_atl08(
                 ground_track=ground_track,
                 filepath=filepath,
                 variables_to_include=gt_variables_to_include,
+                variables_to_check_all_null=gt_variables_to_check_all_null,
             )
             gdfs.append(gdf)
-        except IceSatMissingDataError:
+        except IceSat2MissingDataError:
             continue
 
     if not gdfs:
         msg = f"Found no valid ground track data for {filepath}"
-        raise IceSatMissingDataError(msg)
+        raise IceSat2MissingDataError(msg)
 
     combined_gdf = pd.concat(gdfs)
     combined_gdf.attrs["source_filename"] = filepath.name