+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   [NDS-10](https://bugs.earthdata.nasa.gov/browse/NDS-10) to points by default.
   Allow custom sets of variables via the `gt_variables_to_include` kwarg to
   `read_points_from_atl08`
+- Drop points where, by default, `canopy/h_canopy` and `terrain/h_te_best_fit`
+  are null.
 
 # v0.2.0 (2026-05-08)